chore: move old GetContentTest file to same namespace Doppler.HtmlEditorApi
